refactor: extract advisor logic to service and fix access controls
- Backend: Moved heavy DB execution (transfer and sanitization) from groupAdvisor controller to groupService, implementing a skinny controller architecture.
- Security: Updated RBAC on POST /:groupId/advisor/transfer to allow both 'admin' and 'coordinator' roles.
- Frontend: Fixed broken relative import path in AdvisorAssociationPanel to prevent build failures.

Closes #72
